+++ tests/model_tests/diffusion/conftest.py
@@ -2,7 +2,7 @@ from shutil import rmtree
 
 import pytest
 
-from tests.tiny_models.model_settings import DIFFUSION_TEST_SETTINGS
+from tests.model_tests.diffusion.model_settings import DIFFUSION_TEST_SETTINGS
 
 
 @pytest.fixture(scope="session")
