+++ tests/model_tests/diffusion/task_runners.py
@@ -17,6 +17,9 @@ IMAGE_DIMS = (512, 512)
 HEIGHT, WIDTH = IMAGE_DIMS
 INPUT_IMAGE = Image.new("RGB", IMAGE_DIMS)
 
+# TODO: Add multi-output (n>1) and deterministic (same seed → same output)
+# task runners to cover cases currently only tested in e2e expansion tests.
+
 # Offline sampling params
 IMAGE_GEN_SAMPLING_PARAMS = OmniDiffusionSamplingParams(
     num_inference_steps=4,
